+++ src/insta/server/sql_stats.h
@@ -10,7 +10,7 @@
 
 #include <insta/server/ddnet_db_utils/ddnet_db_utils.h>
 #include <insta/server/extra_columns.h>
-#include <insta/server/generated/mode_account.h>
+#include <insta/server/todo_codegen/mode_account.h>
 #include <insta/server/sql_stats_player.h>
 
 struct ISqlData;
